# faster wino compile by catting consts across data expand dim