Significantly reduces the number of shapetrackers. for now, there is a bug where the gfactors get unfused into 2 kernels instead of 1. but it's way faster.

the idea is as follows:

```
each column of wino transform matrix mat is multiplied by only one element of the input data vector v

before, we have:
Tensor.stack([sum([mat[i][j] * v[i]] for j in range(len(mat[0]))]) for i in range(len(mat))])

this accesses v[i] once for each element of mat, times the size of the output stack, due the pad and add implementation of stack. However, if we first stacked the consts:

sum( [Tensor.stack(mat[i][j] for i in range(len(mat[i]))) * v[j] for j in range(len(mat))])

we can save a factor of len(mat[0]) for each wino dim.

the core idea is: Tensor.stack(mat[i][j] * v[i] for j in range(...)) is slower than Tensor.stack(mat[i][j] for j in range(...)) * v[i]

```

Before:
```
running conv:  45.63 ms
scheduling:  99.59 ms
linearize 1 with  752 ops: 150.95 ms
 297 shapetrackers with max 1 views
linearize 3 with 1495 ops: 445.40 ms
 592 shapetrackers with max 1 views
linearize 4 with    5 ops:   1.67 ms
   3 shapetrackers with max 1 views
linearize 5 with  859 ops: 370.84 ms
 360 shapetrackers with max 2 views

```
After:
```
test_speed (__main__.TestWinograd) ... running conv:  23.79 ms
scheduling:   6.55 ms
linearize 1 with   90 ops:  23.58 ms
  36 shapetrackers with max 1 views
linearize 3 with  256 ops:  84.64 ms
  89 shapetrackers with max 1 views
linearize 4 with    5 ops:   1.61 ms
   3 shapetrackers with max 1 views
linearize 5 with  210 ops: 113.04 ms
  67 shapetrackers with max 2 views
ok
```