@chaosagent 2024-02-01
before, backward fusions were broken. now,

```
WINO=1 BIG=1 DEBUG=0 python test/test_speed_v_torch.py TestBigSpeed.test_large_conv_3x3
conv bs:  4 chans:128 -> 128 k:3            233.15 ms (   39.52 GFLOPS     1.61 GB/s) in torch,    0.61 ms (15086.54 GFLOPS   614.68 GB/s) in tinygrad,    0.00x faster    9215.28 MOPS   375.46 MB

0 17861.64 ms run, 17860.30 ms python,    1.35 ms HIP, 1200.00 loss, 0.000015 LR, 0.42 GB used,    197.28 GFLOPS,   3523.74 GOPS
  1 1062.85 ms run, 1062.30 ms python,    0.55 ms HIP, 1202.00 loss, 0.000030 LR, 2.84 GB used,   3313.63 GFLOPS,   3521.88 GOPS
  2   31.17 ms run,    2.29 ms python,   28.88 ms HIP, 1190.00 loss, 0.000045 LR, 2.85 GB used, 112975.72 GFLOPS,   3521.88 GOPS
315   29.45 ms run,    1.63 ms python,   27.82 ms HIP,  614.00 loss, 0.003063 LR, 2.85 GB used, 119585.80 GFLOPS,   3521.88 GOPS
316   29.55 ms run,    1.65 ms python,   27.90 ms HIP,  601.50 loss, 0.003059 LR, 2.85 GB used, 119181.20 GFLOPS,   3521.88 GOPS
```

compile is not 8s anymore, but both forward and backward wino are faster like this.