@chaosagent 2024-02-01
the big messy lines is because pushing expands is not reliable :(