@chaosagent 2024-02-01
not sure why i don't get sub 30ms steps anymore, maybe perf state ghost, lol

```
  0 19810.94 ms run, 19809.19 ms python,    1.74 ms HIP, 1200.00 loss, 0.000015 LR, 0.42 GB used,    177.87 GFLOPS,   3523.74 GOPS
  1  593.52 ms run,  592.89 ms python,    0.63 ms HIP, 1202.00 loss, 0.000030 LR, 2.84 GB used,   5933.88 GFLOPS,   3521.88 GOPS
  2   32.63 ms run,    2.27 ms python,   30.36 ms HIP, 1190.00 loss, 0.000045 LR, 2.85 GB used, 107922.67 GFLOPS,   3521.88 GOPS
  3   32.14 ms run,    1.62 ms python,   30.52 ms HIP, 1175.00 loss, 0.000060 LR, 2.85 GB used, 109574.69 GFLOPS,   3521.88 GOPS
  4   32.14 ms run,    1.61 ms python,   30.52 ms HIP, 1160.00 loss, 0.000075 LR, 2.85 GB used, 109584.95 GFLOPS,   3521.88 GOPS
  5   32.10 ms run,    1.57 ms python,   30.54 ms HIP, 1160.00 loss, 0.000090 LR, 2.85 GB used, 109712.59 GFLOPS,   3521.88 GOPS
  6   32.18 ms run,    1.62 ms python,   30.57 ms HIP, 1152.00 loss, 0.000105 LR, 2.85 GB used, 109439.58 GFLOPS,   3521.88 GOPS

test_speed (__main__.TestWinograd) ... running conv:  24.06 ms
scheduling:   7.05 ms
linearize 1 with   90 ops:  23.68 ms
  36 shapetrackers with max 1 views
linearize 3 with  256 ops:  85.54 ms
  89 shapetrackers with max 1 views
linearize 4 with    5 ops:   1.63 ms
   3 shapetrackers with max 1 views
linearize 5 with  210 ops: 107.30 ms
  67 shapetrackers with max 2 views
ok

eval     9620/10240 93.95%,    0.40 val_loss STEP=1000 (in 1789.69 ms)

real    1m6.546s
user    1m0.729s
sys     0m11.151s


```