@chaosagent 2024-02-01
with BEAM=2, BS=512 on 7950x:

```
997   18.50 ms run,    0.89 ms python,   17.61 ms HIP,  491.50 loss, 0.000094 LR, 2.85 GB used, 190338.34 GFLOPS,   3521.88 GOPS
998   18.65 ms run,    0.95 ms python,   17.70 ms HIP,  500.75 loss, 0.000089 LR, 2.85 GB used, 188829.70 GFLOPS,   3521.88 GOPS
```