@chaosagent 2024-02-02
tiny4 (epyc 7352)

```
(venv) chaos@tiny4:~/tinygrad$ time HIP_VISIBLE_DEVICES=3 LATEBEAM=8 DEBUG=0 BS=512 HALF=1 LATEWINO=1 python examples/hlb_cifar10.py
shuffling training dataset in 897.93 ms (epoch=0)
  0 23151.31 ms run, 23149.27 ms python,    2.04 ms HIP, 1200.00 loss, 0.000015 LR, 0.42 GB used,    152.20 GFLOPS,   3523.74 GOPS
  1 1019.08 ms run, 1018.52 ms python,    0.56 ms HIP, 1202.00 loss, 0.000030 LR, 2.84 GB used,   3455.92 GFLOPS,   3521.88 GOPS
  2   19.32 ms run,    2.24 ms python,   17.08 ms HIP, 1190.00 loss, 0.000045 LR, 2.85 GB used, 182316.04 GFLOPS,   3521.88 GOPS
  3   18.69 ms run,    1.69 ms python,   17.00 ms HIP, 1175.00 loss, 0.000060 LR, 2.85 GB used, 188459.13 GFLOPS,   3521.88 GOPS
  4   18.50 ms run,    1.63 ms python,   16.87 ms HIP, 1160.00 loss, 0.000075 LR, 2.85 GB used, 190366.80 GFLOPS,   3521.88 GOPS
  5   18.38 ms run,    1.64 ms python,   16.74 ms HIP, 1160.00 loss, 0.000090 LR, 2.85 GB used, 191609.22 GFLOPS,   3521.88 GOPS
  6   18.41 ms run,    1.64 ms python,   16.76 ms HIP, 1152.00 loss, 0.000105 LR, 2.85 GB used, 191319.24 GFLOPS,   3521.88 GOPS
  7   18.31 ms run,    1.65 ms python,   16.66 ms HIP, 1172.00 loss, 0.000120 LR, 2.85 GB used, 192327.34 GFLOPS,   3521.88 GOPS
  8   18.27 ms run,    1.64 ms python,   16.64 ms HIP, 1147.00 loss, 0.000134 LR, 2.85 GB used, 192721.05 GFLOPS,   3521.88 GOPS
  9   18.34 ms run,    1.64 ms python,   16.70 ms HIP, 1137.00 loss, 0.000149 LR, 2.85 GB used, 192026.60 GFLOPS,   3521.88 GOPS

...

978   18.74 ms run,    1.63 ms python,   17.12 ms HIP,  504.50 loss, 0.000178 LR, 2.85 GB used, 187899.60 GFLOPS,   3521.88 GOPS
979   18.57 ms run,    1.61 ms python,   16.96 ms HIP,  502.50 loss, 0.000173 LR, 2.85 GB used, 189642.58 GFLOPS,   3521.88 GOPS
shuffling training dataset in 1138.67 ms (epoch=10)
980 1158.06 ms run, 1141.12 ms python,   16.95 ms HIP,  495.25 loss, 0.000168 LR, 2.85 GB used,   3041.67 GFLOPS,   3522.44 GOPS
981   18.77 ms run,    1.74 ms python,   17.03 ms HIP,  495.00 loss, 0.000164 LR, 2.85 GB used, 187664.00 GFLOPS,   3521.88 GOPS
982   18.48 ms run,    1.65 ms python,   16.83 ms HIP,  494.75 loss, 0.000160 LR, 2.85 GB used, 190551.68 GFLOPS,   3521.88 GOPS
983   18.41 ms run,    1.62 ms python,   16.78 ms HIP,  492.25 loss, 0.000155 LR, 2.85 GB used, 191352.18 GFLOPS,   3521.88 GOPS
984   18.34 ms run,    1.61 ms python,   16.73 ms HIP,  488.00 loss, 0.000151 LR, 2.85 GB used, 192056.22 GFLOPS,   3521.88 GOPS
985   18.24 ms run,    1.63 ms python,   16.61 ms HIP,  495.00 loss, 0.000146 LR, 2.85 GB used, 193061.65 GFLOPS,   3521.88 GOPS
986   18.27 ms run,    1.67 ms python,   16.60 ms HIP,  503.25 loss, 0.000142 LR, 2.85 GB used, 192797.44 GFLOPS,   3521.88 GOPS
987   18.47 ms run,    1.64 ms python,   16.83 ms HIP,  492.50 loss, 0.000138 LR, 2.85 GB used, 190717.60 GFLOPS,   3521.88 GOPS
988   18.53 ms run,    1.60 ms python,   16.93 ms HIP,  499.75 loss, 0.000133 LR, 2.85 GB used, 190028.56 GFLOPS,   3521.88 GOPS
989   18.30 ms run,    1.66 ms python,   16.64 ms HIP,  495.50 loss, 0.000128 LR, 2.85 GB used, 192424.96 GFLOPS,   3521.88 GOPS
990   18.46 ms run,    1.61 ms python,   16.85 ms HIP,  496.00 loss, 0.000125 LR, 2.85 GB used, 190757.47 GFLOPS,   3521.88 GOPS
991   18.50 ms run,    1.61 ms python,   16.89 ms HIP,  497.75 loss, 0.000120 LR, 2.85 GB used, 190366.90 GFLOPS,   3521.88 GOPS
992   18.43 ms run,    1.65 ms python,   16.78 ms HIP,  494.50 loss, 0.000115 LR, 2.85 GB used, 191089.81 GFLOPS,   3521.88 GOPS
993   18.47 ms run,    1.60 ms python,   16.88 ms HIP,  491.75 loss, 0.000112 LR, 2.85 GB used, 190633.57 GFLOPS,   3521.88 GOPS
994   18.74 ms run,    1.63 ms python,   17.12 ms HIP,  492.25 loss, 0.000107 LR, 2.85 GB used, 187911.12 GFLOPS,   3521.88 GOPS
995   18.64 ms run,    1.65 ms python,   16.99 ms HIP,  496.25 loss, 0.000102 LR, 2.85 GB used, 188972.63 GFLOPS,   3521.88 GOPS
996   18.66 ms run,    1.61 ms python,   17.05 ms HIP,  497.75 loss, 0.000099 LR, 2.85 GB used, 188722.70 GFLOPS,   3521.88 GOPS
997   18.80 ms run,    1.62 ms python,   17.18 ms HIP,  495.75 loss, 0.000094 LR, 2.85 GB used, 187338.77 GFLOPS,   3521.88 GOPS
998   18.85 ms run,    1.62 ms python,   17.23 ms HIP,  498.25 loss, 0.000089 LR, 2.85 GB used, 186840.75 GFLOPS,   3521.88 GOPS
999   18.75 ms run,    1.66 ms python,   17.10 ms HIP,  495.00 loss, 0.000086 LR, 2.85 GB used, 187808.61 GFLOPS,   3521.88 GOPS
shuffling test dataset in 51.52 ms (epoch=0)
eval     9615/10240 93.90%,    0.40 val_loss STEP=1000 (in 994.26 ms)

real    1m0.051s
user    0m54.165s
sys     0m11.222s

```