@chenyuxyz 2024-02-02
will take a closer look later. i think it breaks multi gpu cifar `time LATEWINO=1 HALF=1 GPUS=2 STEPS=5 python3 examples/hlb_cifar10.py`. device was set to default for `Tensor(float(m[k]))`